changed food IV-rating to double. create foodRating dummy data for DB

+++ Backend/business/src/main/resources/foodSchema.sql
@@ -8,7 +8,7 @@ CREATE TABLE food(
 	price VARCHAR(20),
 	category varchar(80),
 	located_at INT NOT NULL,
-	rating INT DEFAULT 0,
+	rating DOUBLE DEFAULT 0,
 	PRIMARY KEY(food_id),
 	FOREIGN KEY(located_at) REFERENCES restaurant(restaurant_id)
 );
